requirements.txt
asgiref==3.6.0
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cssselect2==0.7.0
Django==4.2.1
django-formset-js==0.5.0
django-jquery-js==3.1.1
django-taggit==4.0.0
django-weasyprint==2.2.0
django-widget-tweaks==1.4.12
fonttools==4.39.3
html5lib==1.1
idna==3.4
Pillow==9.5.0
psycopg2==2.9.6
pycparser==2.21
pydyf==0.6.0
pyphen==0.14.0
python-decouple==3.8
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
tzdata==2023.3
urllib3==2.0.2
weasyprint==58.1
webencodings==0.5.1
zopfli==0.2.2
